# Add new method bumpLru in order to get rid of redundant operations