@kibertoad 2023-04-01
@avoidwork Can you elaborate a little bit? This preserves all the useful behaviours while improving performance.