@avoidwork 2023-04-02
The code path was tuned for maximizing performance across all ops. What's the motivation to change it? This is very similiar to an earlier version of the code.