@avoidwork 2023-04-02
I mean, is there a legitimate performance gain that I'm not understanding? Numbers would help if so. The benchmark that's included is just a very light touch to see if something's going wrong; bench-lru repo may still be useful but I think it's outdate.