@kibertoad 2023-04-02
@avoidwork This completely avoids calling `set` on every get, which is a legitimate performance gain in my book. And it makes overall behaviour less fragile and more predictable. I'll open separate PR with a much smaller change which does not improve perf, but addresses a very important expiration gap, just to reduce cognitive load of this PR.