@avoidwork 2023-04-02
Did you come to that conclusion with IR logs? This has been the design since the initial commit, unless you're going at this with tooling to validate your assumption it's not going to be accepted.

https://github.com/avoidwork/tiny-lru/blob/0c64c8acb7b18bd565e3ac9d5b952620ecae4df0/src/lru.js#L54