@kibertoad 2023-04-02
@avoidwork Which assumption? It's a matter of fact that every GET operation invokes SET operation. If that is by design, and you don't want to avoid that, and are OK with perf penalty that you are getting there, that's fine, I'm not going to convince you.

Please take a look at https://github.com/avoidwork/tiny-lru/pull/75/files, however, I strongly believe it makes behaviour more correct in all circumstances.