@avoidwork 2023-04-02
What i'm trying to say is the compilers used to inline it, why would they stop? I tried looking at things with deoptigate yesterday but the logs are too different now; however there's more than 1 deopt that used to exist which is unfortunate.

I don't know if everyone is testing their proposed changes to the same depth I was, but this should be somewhat provable with the PR on bench-lru; it's pinned to 7.x.x atm, so just do 3 runs of that, 10.3.0, and your PR, without the other caches to see how it performs.

Running the bundled benchmark.js through deoptigate will get you the logs to look at.