@avoidwork 2023-04-02
That's the key point, that is easily missed. When there was 1 deopt, it was due to mega shape change on a key value inside of set(), and there was no deopt on get() calling set() until the significant amount of shape changes ... your PR to use a Map changes that entirely.