@avoidwork 2023-04-02
The Map made everything worse, see https://github.com/avoidwork/tiny-lru/pull/76